refactor: rename index to artworksidebar2 and pull BelowTheFoldArtworkDetails out
